refactor(lane_change): move for lane change module to interface (#3536)
* refactor(lane_change): move for lane change module to interface


* fix build error


* fix conflict

---------
